\name{drd3}
\docType{data}
\alias{drd3}
\title{An example dataset}
\description{relative DRD3 mRNA levels in blood of 37 schizophrenia cases and 37 healthy controls}
\usage{data(drd3)}
\format{A data.frame containing 74 observations from two groups.}
\source{Converging evidence implicates the dopamine D3 receptor gene in vulnerability to schizophrenia. 
 Am J Med Genet Part B. 2011, 156(5):613-9}
\references{
Zhang, F. et al. \emph{Converging evidence implicates the dopamine D3 receptor gene in vulnerability to schizophrenia}. Am J Med Genet Part B. 2011, 156(5):613-9.
}
\keyword{datasets}