\docType{package}
\name{plot2groups-package}
\alias{plot2groups-package}
\title{The plot2groups Package}
\description{
  Plot scatter points for two groups of values.
}
\details{
  \tabular{ll}{ Package: \tab plot2groups\cr Type: \tab
  Package\cr Version: \tab 0.10\cr Date: \tab 2014-03-15\cr
  Depends: \tab R (>= 2.15.0)\cr Encoding: \tab UTF-8\cr
  License: \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr }

  A function to plot scatter points for two groups of
  values. It simultaneously calculates a P value of t-test
  and incorporate the P value into the plot. Its advantage
  is jittering the adjacent points side by side to avoid or
  reduce overlapping in the plot.
}
\author{
  Fuquan Zhang
}
\keyword{package}

